# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Believer Backend API
  version: 1.8.0
  description: Active REST API contract for auth, account recovery, account-governance settings actions, business listings, mosques, backend-owned mosque prayer times, mosque page content, mosque image uploads, bookmarks, notifications, and services.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request payload
        details:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          type: object
          example: {}
...
    User:
      type: object
      required: [id, fullName, email, role]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [community, admin, super_admin]
    TokenPair:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    AuthPayload:
      type: object
      required: [user, tokens]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/TokenPair'
    AuthEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/AuthPayload'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    RefreshPayload:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RefreshEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/RefreshPayload'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    UserEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/User'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    SuccessFlagEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [success]
          properties:
            success:
              type: boolean
              example: true
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    SuccessMessageEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [success, message]
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: If an account exists for that email, a password reset link has been sent.
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    SignupRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        accountType:
          type: string
          enum: [community, admin]
          default: community
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 512
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UserUpdateRequest:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 8
          maxLength: 128
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Create account
      description: Community signup is the default. Mosque admin signup is also supported through the same full-name, email, and password flow by setting `accountType=admin`. After a successful signup, the backend attempts a lightweight welcome email without blocking account creation if delivery fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: Returns the same success payload whether or not the email belongs to an active account. Requires configured transactional email and sends a one-time reset link through the backend email service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Email provider failed to deliver the reset email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Password reset email is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with a one-time token
      description: Consumes a hashed single-use reset token, updates the password, and revokes previously issued refresh-token state for the account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessFlagEnvelope'
        '400':
          description: Invalid or expired reset token, or malformed payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          description: Invalid profile payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed and tokens rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          description: Invalid payload or wrong current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logout success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessFlagEnvelope'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...